+++ backend/csvimporter/src/main/java/com/simra/app/csvimporter/controller/MonitorService.java
@@ -1,11 +1,6 @@
 package com.simra.app.csvimporter.controller;
 
 
-import java.io.IOException;
-
 public interface MonitorService {
-    public void init() throws IOException;
-
-    public void startRecursiveWatcher();
-
+    public void startRecursiveWatcher() throws Exception;
 }
